+++ src/ll/xist/ns/xml.py
@@ -42,8 +42,14 @@ class XML(xsc.ProcInst):
 	"""
 	xmlname = "xml"
 
-	def __init__(self, version="1.0", encoding="utf-8"):
-		xsc.ProcInst.__init__(self, u'version="%s" encoding="%s"' % (version, encoding))
+	def __init__(self, version="1.0", encoding="utf-8", standalone=None):
+		v = []
+		v.append(u'version="%s"' % version) # According to http://www.w3.org/TR/2000/REC-xml-20001006#NT-XMLDecl version is required
+		if encoding is not None:
+			v.append(u'encoding="%s"' % encoding)
+		if standalone is not None:
+			v.append(u'standalone="%s"' % ("yes" if standalone else "no"))
+		xsc.ProcInst.__init__(self, u" ".join(v))
 
 
 class XMLStyleSheet(xsc.ProcInst):